Refactoring Database code a bit
git-svn-id: svn+ssh://localhost/home/svn/musador/trunk@136 abf014a6-d86a-7740-8175-6a5112e866a6

+++ src/Database/test/Database.h
@@ -12,7 +12,7 @@ class DatabaseTest : public CxxTest::TestSuite
 
 public:
 
-    unsigned int entId;
+    __int64 entId;
 
     DatabaseTest()
     {
